Add README to download jdk1.7

+++ docker/build.sh
@@ -1,3 +1,5 @@
-cp ~/jdk-7u80-linux-x64.tar.gz .
+if [ -f ~/jdk-7u80-linux-x64.tar.gz ]; then
+	cp ~/jdk-7u80-linux-x64.tar.gz .
+fi
 
-docker build -t test-docker .
+docker build -t docker-demo .
